+++ teacher/delete_quiz.php
@@ -1,12 +1,17 @@
 <?php
-include('../dbcon.php');
-if (isset($_POST['backup_delete'])){
-$id=$_POST['selector'];
-$N = count($id);
-for($i=0; $i < $N; $i++)
-{
-	$result = mysqli_query($con,"DELETE FROM quiz where quiz_id='$id[$i]'");
+include('../admin/dbcon.php');
+if (isset($_POST['backup_delete'])) {
+	if ($_POST['selector']!=null){
+		$id = $_POST['selector'];
+		$N = count($id);
+		for ($i = 0; $i < $N; $i++) {
+			$result = mysqli_query($con,"DELETE FROM quiz where quiz_id='$id[$i]'");
+		}
+		header("location: teacher_quiz.php");
+	}else{
+		header("location: teacher_quiz.php");
+	}
+	
+}else{
+	header("location: teacher_quiz.php");
 }
-header("location: teacher_quiz.php");
-}
-?>
